move super and yield from to a legacy plugin
